Fixed Failures in Schedule and Lock Specs
With the use of the new ORM layer we can no longer guarentee that
a created job will evaluate as '==' to a job retrieved from the
ORM.  This is a property that as of this writing is fairly exclusive
to DataMapper.  This is a temperary fix that will only work with
ORMs very simular to DataMapper.

TODO: 'Update.current.get(u.id)' in schedule_spec.rb needs to be ORM
agnostic.

In changing from a SQL spesific format inherited from the venerable
delayed_job to a more ORM agnostic situation, the consept of locking
a record  before running it moved from the aplication layer to the
ORM with the #lock_next method.  It is up to each driver to deturmine
how best to lock a record in the presence of multiple simultanious
worker clients.  MongoDB for example has facilities to find and update
a record atomicly and can forgo the "select several and lock random"
method compleatly.

Note: Changes in the worker model have obsoleted a number of the
Worker specs.  Until noted otherwise in the commit tree,
'test/fork.rb' and 'test/thread.rb' are the devinative tests for
these classes.