Parameter Substitution Tested
There are three special parameters that when passed will be replaced
by special values.  __job__ will be replaced by the job that chained-in
to this job, __params__ will be replaced by by a user defined hash
stored on the chain request if it is present, __self__ will always
be the job that actually called the method.  If this is not a chained
method, __job__ and __params__ will be replaced with nil.

These methods are intended primarly to aid in library chained methods
like 'reschedule' or 'log_error' (neither yet implimented), but can be
used to good effect by advanced updater aware client methods as well.