Fixed Errors in fork_worker_instance_spec
There were changes maide based on the test/fork.rb runs.  One changed
the signature of #smoke_pipe to take an IOStream instead of an array
from Pipe.new.  The other was to block heartbeat if continue was false
sence the inode that we are using as a shared state may have been closed.

These changes bring the test suite in-line with the working code.