Client can set logger Directly
A new option in setup#initialize allows a client to pass in options
which overide those in the configuration file, and adds a :logger
option which directly sets the logger.  With this option :log_file
is ignored.

A temperary debugging feature was added which checks the validity
of the @orm iVar in class scope of ForkWorker.  There is a rare
bug that causes this valuse to be changed to and array with the
single entry of the set ORM class while exicuting a worker instance.

I estimate that this occurse about 1 time in 5000 jobs and may only
effect the first instance forked.